+++ apt/apt-pkg/rpm/rpmsrcrecords.cc
@@ -117,7 +117,7 @@ bool rpmSrcRecordParser::Step()
    return true;
 }
 
-bool rpmSrcRecordParser::Jump(unsigned long Off)
+bool rpmSrcRecordParser::Jump(unsigned long long Off)
 {
    if (!Handler->Jump(Off))
        return false;
@@ -232,7 +232,7 @@ string rpmSrcRecordParser::Changelog() const
    return rval;
 }
 
-unsigned long rpmSrcRecordParser::Offset() 
+unsigned long long rpmSrcRecordParser::Offset()
 {
     return Handler->Offset();
 }
